2026.03.25 + Fix everything

+++ 2026.03.20/8./main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 
 int main() {
+
     int elso;
     int masodik;
 
@@ -10,7 +11,7 @@ int main() {
     std::cout << "Masodik nepszamlalas: ";
     std::cin >> masodik;
 
-    std::cout << masodik / elso * 100 - 100 << "% volt a novekedes." << std::endl;
+    std::cout << masodik / (double)elso * 100 - 100 << "% volt a novekedes." << std::endl;
 
     return 0;
 }
